+++ OUDS/Core/Themes/Orange/Sources/OrangeTheme.swift
@@ -115,11 +115,14 @@ open class OrangeTheme: OUDSTheme, @unchecked Sendable {
     ///    - inputText: All component tokens for input text
     ///    - link: All component tokens for link
     ///    - listItem: All component tokens for list item
+    ///    - pinCodeInput: All component tokens for pin code input
+    ///    - quantityInput: All component tokens for quantity input
     ///    - radioButton: All component tokens for radio buttons
     ///    - select: All component tokens for select
     ///    - skeleton: All component tokens for skeleton
     ///    - switch: All component tokens for switch
     ///    - tag: All component tokens for tag
+    ///    - tagInput: All component tokens for tag input
     ///    - resourcesBundle: The `Bundle` of the module containing assets to laod like images
     ///    - fontFamily: Set `nil` if system font to use, otherwise use the `FontFamilySemanticToken` you want to apply
     override public init(borders: AllBorderSemanticTokensProvider? = nil,
@@ -143,11 +146,14 @@ open class OrangeTheme: OUDSTheme, @unchecked Sendable {
                          inputText: AllInputTextComponentTokensProvider? = nil,
                          listItem: AllListItemComponentTokensProvider? = nil,
                          link: AllLinkComponentTokensProvider? = nil,
+                         pinCodeInput: AllPinCodeInputComponentTokensProvider? = nil,
+                         quantityInput: AllQuantityInputComponentTokensProvider? = nil,
                          radioButton: AllRadioButtonComponentTokensProvider? = nil,
                          select: AllSelectComponentTokensProvider? = nil,
                          skeleton: AllSkeletonComponentTokensProvider? = nil,
                          switch: AllSwitchComponentTokensProvider? = nil,
                          tag: AllTagComponentTokensProvider? = nil,
+                         tagInput: AllTagInputComponentTokensProvider? = nil,
                          resourcesBundle: Bundle = Bundle.OrangeTheme,
                          fontFamily: FontFamilySemanticToken? = nil)
     {
@@ -174,11 +180,14 @@ open class OrangeTheme: OUDSTheme, @unchecked Sendable {
         let inputText = (inputText ?? OrangeThemeInputTextComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces))
         let link = (link ?? OrangeThemeLinkComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces))
         let listItem = (listItem ?? OrangeThemeListItemComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces))
+        let pinCodeInput = (pinCodeInput ?? OrangeThemePinCodeInputComponentTokensProvider(sizes: sizes, spaces: spaces))
+        let quantityInput = (quantityInput ?? OrangeThemeQuantityInputComponentTokensProvider(sizes: sizes, spaces: spaces))
         let radioButton = (radioButton ?? OrangeThemeRadioButtonComponentTokensProvider(sizes: sizes, borders: borders))
         let select = (select ?? OrangeThemeSelectComponentTokensProvider(sizes: sizes, colors: colors, spaces: spaces))
         let skeleton = (skeleton ?? OrangeThemeSkeletonComponentTokensProvider(colors: colors))
         let `switch` = (`switch` ?? OrangeThemeSwitchComponentTokensProvider(sizes: sizes, borders: borders, colors: colors, spaces: spaces, opacities: opacities, dimensions: dimensions))
         let tag = (tag ?? OrangeThemeTagComponentTokensProvider(sizes: sizes, borders: borders, spaces: spaces, dimensions: dimensions))
+        let tagInput = (tagInput ?? OrangeThemeTagInputComponentTokensProvider(borders: borders, colors: colors))
 
         super.init(borders: borders,
                    colors: colors,
@@ -201,11 +210,14 @@ open class OrangeTheme: OUDSTheme, @unchecked Sendable {
                    inputText: inputText,
                    listItem: listItem,
                    link: link,
+                   pinCodeInput: pinCodeInput,
+                   quantityInput: quantityInput,
                    radioButton: radioButton,
                    select: select,
                    skeleton: skeleton,
                    switch: `switch`,
                    tag: tag,
+                   tagInput: tagInput,
                    resourcesBundle: Bundle.OrangeTheme,
                    fontFamily: fontFamily)
     }