+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Quantity input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Tag input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Pin code input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Color charts for Orange Business Tools theme